fix version detection. use pusre lwrp
