+++ resources/npm.rb
@@ -23,7 +23,7 @@ actions :install, :upgrade, :remove
 attribute :package_name,  kind_of: String, name_attribute: true
 attribute :version,       kind_of: String
 attribute :source,        kind_of: String
-attribute :node_version,  kind_of: String
+attribute :node_version,  kind_of: String, required: true
 attribute :response_file, kind_of: String
 
 def initialize(*args)
